backend/requirements.txt
langchain-core==0.3.76
langgraph==0.6.8
langchain-cerebras==0.5.0
cerebras-cloud-sdk==1.0.0
langsmith==0.4.31
python-dotenv==1.0.1
pydantic==2.11.9
httpx==0.27.2
fastapi==0.118.0
uvicorn==0.30.6
gunicorn==23.0.0
pymongo==4.9
motor==3.6.0
urllib3==1.26.18
six==1.16.0
certifi>=2024.2.2
